+++ src/page/Issues.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import { useQuery} from '@apollo/client';
 import { useParams } from 'react-router-dom';
-import {
-    Box, Button, Flex, Heading, Text,
-    Card, CardBody, Stack, Divider
-} from '@chakra-ui/react';
-import { FaGithub } from 'react-icons/fa';
+import { Card, CardBody } from '@chakra-ui/react';
 import { GET_ISSUES } from '../query/SearchIssues';
 import Loading from '../components/Loading'
 import PageTitle from '../components/PageTitle'
+import IssuesHeader from '../components/IssuesHeader'
+import IssuesList from '../components/IssuesList';
 
 const Issues: React.FC = () => {
     const params = useParams();
@@ -19,9 +17,6 @@ const Issues: React.FC = () => {
         }
     });
 
-    const PageBack = () => {
-        window.history.back()
-    }
 
     if (loading) return <Loading />;
     if (error) return <p>Error :{error.message}</p>;
@@ -33,33 +28,8 @@ const Issues: React.FC = () => {
             {data && (
                 <Card>
                     <CardBody>
-                        <Button onClick={PageBack}>← back to search</Button>
-                        <Flex my="8">
-                            <Heading size='sm' mr="2">
-                                Latest issues
-                            </Heading>
-                            <Text fontSize='xs' color="gray.400">
-                                on {data.node.owner.login}/{data.node.name}
-                            </Text>
-                        </Flex>
-                        <Stack>
-                            {data.node.issues.edges.map(( edge: any, index: number ) => (
-                                <Box key={index}>
-                                    <Box p="3">
-                                        <a href={edge.node.url} target="_blank" rel="noopener noreferrer">
-                                            <Flex>
-                                                <Text fontSize='md' color="gray.600" mr="3">
-                                                    {edge.node.title}
-                                                </Text>
-                                                {/* <FiExternalLink /> */}
-                                                <FaGithub />
-                                            </Flex>
-                                        </a>
-                                    </Box>
-                                    <Divider />
-                                </Box>
-                            ))}
-                        </Stack>
+                        <IssuesHeader data={data} />
+                        <IssuesList data={data} />
                     </CardBody>
                 </Card>
             )}
